html/browsers/browsing-the-web/back-forward-cache/events-freeze.html

<!DOCTYPE HTML>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/html/cross-origin-embedder-policy/credentialless/resources/dispatcher.js"></script>
<script src="resources/helper.sub.js"></script>
<script>
// https://github.com/whatwg/html/pull/5889
promise_test(async t => {
  const idA = token();

  window.open(
    executorPath + idA + '&events=visibilitychange,pagehide,pageshow,load,freeze,resume',
    '_blank', 'noopener');

  await send(idA, `
      prepareNavigation();
      location.href = '${originCrossSite + backPath}';
  `);

  await assert_bfcached(idA);

  assert_array_equals(
    await evalOn(idA, `getRecordedEvents()`),
    [
      'window.load',
      'window.pageshow',
      'window.pagehide.persisted',
      'document.visibilitychange.hidden',
      'window.visibilitychange.hidden',
      // document.freeze/resume are not fired on Firefox/Safari.
      'document.freeze',
      'document.resume',
      'document.visibilitychange.visible',
      'window.visibilitychange.visible',
      'window.pageshow.persisted'
    ]);

}, 'Events fired (document.freeze/resume)');
</script>
